All or nothing

+++ ClientApp/ModernEncryption/Service/UserService.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Diagnostics;
 using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
@@ -6,6 +7,7 @@ using ModernEncryption.BusinessLogic.UserManagement;
 using ModernEncryption.Interfaces;
 using ModernEncryption.Model;
 using Newtonsoft.Json;
+using Org.BouncyCastle.Crypto;
 using Plugin.SecureStorage;
 
 namespace ModernEncryption.Service
@@ -33,12 +35,16 @@ namespace ModernEncryption.Service
 
         public async Task<bool> CreateUser(User user)
         {
+            var validateEmail = GetUser(user.Email).Result;
+            if (validateEmail == null) return false;
             var uri = new Uri(string.Format(Constants.RestUrlNewUser));
             var json = JsonConvert.SerializeObject(user);
             var content = new StringContent(json, Encoding.UTF8, "application/json");
+            Debug.WriteLine(json);
 
             HttpResponseMessage response = null;
             response = await _client.PostAsync(uri, content);
+            CrossSecureStorage.Current.SetValue("RegistrationProcess", "1");
 
             var voucherCode = new VoucherCode(user);
             voucherCode.SendVoucherCode();
@@ -49,8 +55,11 @@ namespace ModernEncryption.Service
         public bool ValidateVoucherCode(int userVoucher)
         {
             var voucher = Convert.ToInt32(CrossSecureStorage.Current.GetValue("Voucher", "-1"));
+            var voucherTimestamp = Convert.ToInt32(CrossSecureStorage.Current.GetValue("VoucherTimestamp", "-1"));
+            Int32 CurrentVoucherTimestamp = (Int32) (DateTime.UtcNow.Subtract(new DateTime(1970, 1, 1))).TotalSeconds;
+            if (voucherTimestamp > CurrentVoucherTimestamp) return false;
             if (voucher != userVoucher) return false;
-
+            CrossSecureStorage.Current.DeleteKey("VoucherTimestamp");
             CrossSecureStorage.Current.DeleteKey("Voucher");
             return true;
         }
